backend/pyproject.toml
[project]
name = "django-ninja-next"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ninja-jwt>=5.3.5",
    "django>=5.1.5",
    "django-ninja>=1.3.0",
    "djoser>=2.3.1",
    "factory-boy>=3.3.1",
    "pytest-django>=4.9.0",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
    "dj-database-url>=2.3.0",
    "whitenoise>=6.9.0",
    "django-allauth[socialaccount]>=65.10.0",
    "stripe>=12.3.0",
    "click>=8.2.1",
    "uvicorn>=0.35.0",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.12.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    "migrations",
    "__pycache__",
    ".git",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*/settings.py" = ["F405", "F403"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
